+++ src/index.ts
@@ -43,7 +43,7 @@ export * from '@components/editor/editor';
 export * from '@components/code-editor/code-editor';
 export * from '@components/Loader/Loader';
 export * from '@components/async-autocomplete/async-autocomplete';
-export * from '@components/sidebar/side-bar';
+export * from '@components/sidebar/sidebar';
 // Other stuff
 export * from '@src/lib/icons';
 export * from '@src/reports/summary-tile/summary-tile';
